rename: 패키지 이름 Camel case 변경
